requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
gunicorn==21.2.0
requests==2.32.3
python-dotenv==1.0.1
httpz=0.28.1